+++ book/report.go
@@ -4,18 +4,42 @@ import (
 	"fmt"
 )
 
+// Report represents a set of indications about libro's automatic or
+// semi-automatic activities to the end-user that deserve attention or
+// arbitration.
+type Report struct {
+	// Issues collects messages that report events encountered during Book's
+	// processing that deserve end-user attention.
+	Issues []string `json:",omitempty"`
+
+	// SimilarBooks collects alternative Book's metadata that are possibly
+	// better or more complete than actual metada set. `libro`is for some reasons
+	// usually not sure enough whether they are corresponding to exactly the
+	// same book.
+	SimilarBooks []*Book `json:",omitempty"`
+}
+
+// NewReport creates a new empty Report.
+func NewReport() *Report {
+	return &Report{}
+}
+
 // ReportIssue reports a (possible) issue encountered during Book's processing
 // that deserves end-user attention.
-func (b *Book) ReportIssue(format string, a ...interface{}) {
+func (r *Report) ReportIssue(format string, a ...interface{}) {
 	msg := fmt.Sprintf(format, a...)
-	b.ToReview = append(b.ToReview, msg)
+	r.Issues = append(r.Issues, msg)
 	Verbose.Print("warn: " + msg)
 }
 
-// ReportSuggestion reports a (possible) better choice for Book's
-// attribute(s) that can benefit to end-user.
-func (b *Book) ReportSuggestion(format string, a ...interface{}) {
-	msg := fmt.Sprintf(format, a...)
-	b.ToReview = append(b.ToReview, msg)
-	Verbose.Print(msg)
+// ReportSimilarBook reports possible similar Book.
+func (r *Report) ReportSimilarBook(book *Book) {
+	r.SimilarBooks = append(r.SimilarBooks, book)
+	Verbose.Printf("found similar book: %v", book)
+}
+
+// NeedReview returns whether Report contains significant elements for the
+// end-user to consider.
+func (r Report) NeedReview() bool {
+	return len(r.Issues) > 0 || len(r.SimilarBooks) > 0
 }