+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 ## [0.2.1]
+- change Books' metadata completion logic by avoiding using googlebooks search
+  results if ISBN identifiers are not matching. Libro prefers in this case to
+  rely on the end-user choice.
 - add a reporting function that draws end-user attention on possible issues
   encountered during Book's processing.
 - add support to correct reversed Authors name